@@ -26,7 +26,7 @@ import org.keycloak.provider.ServerInfoAwareProviderFactory;
 import org.keycloak.representations.IDToken;
 import org.keycloak.sessions.AuthenticationSessionModel;
 
-import javax.ws.rs.core.Response;
+import jakarta.ws.rs.core.Response;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
